feat: Moved to new util TS file
